requirements.txt

request
dotenv
os